Working login signup and verification for email
need to implement the "and email was sent to freder****@gmail.com" and add the "made a mistake in your email? change it here" Also need to add a reset password along with a edit email. Also need to add a forgot email and two factor authentication and also google authentication. Also need to add a settings tab where the user can change all these settings. Maybe more ideas will come to me as i work on this. Also need to restructure files inside as it's quite messy.